Create reverse_linked_list.py
I looked around and noticed that there's no algorithm that reverse a linked list and returns them. The closest I could find prints them instead of returning them as a function. And so, here it is.

The function only have one input and one output. The function works as follows; it pops data from the beginning of input and append them to the beginning of output.

For tests, I added doctests for short and concise explanation as to how to use the function, and additional tests if the code is run independently. Hopefully, this piece of code and its documentation could be a suggestion for someone that needs to know how to reverse a linked list.